Made AvatarComponent make Profile more modulair. Avatar functionality, include css, etc have now own component
